requirements.txt
appnope==0.1.0
colorlover==0.2.1
decorator==4.0.10
entrypoints==0.2.2
future==0.16.0
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.2
pandas==0.19.1
pexpect==4.2.1
pickleshare==0.7.4
plotly==1.12.9
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
python-highcharts==0.3.1
python-magic==0.4.12
pytz==2016.7
PyYAML==3.12
pyzmq==16.0.0
qtconsole==4.2.1
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
voluptuous==0.9.3
wcwidth==0.1.7
widgetsnbextension==1.2.6
xlrd==1.0.0
